Improve testability and reorg code
